+++ lib/nodo/script.rb
@@ -2,12 +2,14 @@ module Nodo
   class Script
     attr_reader :code
     
-    def initialize(code)
-      @code = code
+    def initialize(code = nil, &block)
+      raise ArgumentError, 'cannot give code when block is given' if code && block
+      @code = code || block
     end
   
     def to_js
-      "#{code}\n"
+      js = code.respond_to?(:call) ? code.call : code
+      "#{js}\n"
     end
   end
 end