+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.6.3
+
+- use Nodo.timeout for all JS calls
+- deferred code generation for scripts and functions
+
 ## 1.6.2
 
 - context-based evaluation